Refactored ExprParser (Created PropertyReferenceParser)
