+++ src/foam/lib/json/AnyParser.java
@@ -18,6 +18,7 @@ public class AnyParser
         new BooleanParser(),
         new LongParser(),
         new DoubleParser(),
-        new ExprParser()));
+        new PropertyReferenceParser(),
+        new FObjectParser()));
   }
 }
